refactor(resolver): rename ScriptAssembliesCache to RecompileGate, add Reset() facade, consolidate define-iteration
The class no longer matches its name: ScriptAssembliesCache implies it
only touches Library/ScriptAssemblies, but it also manipulates
PlayerSettings via RemoveReadyDefine. Both call sites
(UpdatePopupWindow.StartPackageInstall, NuGetResolverMenu.ForceResolve)
also always invoked Wipe() and RemoveReadyDefine() in lockstep —
Reviewer flagged the duplication and the misnomer.

- Rename class+file to RecompileGate (git mv preserves history+meta GUID)
  to match the actual responsibility: gating Unity's next compile to
  include only the resolver itself.

- Add public Reset() facade that does Wipe() + RemoveReadyDefine() so
  callers can't accidentally do half the operation.

- Move EnsureScriptingDefine + TryAddDefine off NuGetDependencyResolver
  onto RecompileGate.EnsureReadyDefine (paired with RemoveReadyDefine).
  Both now share a single ForEachTarget(Func<NamedBuildTarget, bool>)
  helper that iterates BuildTargetGroup values + the distinct
  NamedBuildTarget.Server — the iteration boilerplate that was mirrored
  across both methods is now declared once.

- Trim verbose call-site comment blocks to a single WHY sentence each.

Net: -190 / +178 lines, mostly comment trim + dedupe. NuGetDependencyResolver
loses ~65 lines of editor-build-target plumbing and is now focused
purely on the [InitializeOnLoad] flow.
